Basically started over with how things are represented.
No more class inheritance. I'm going to use more of an
old-school structured programming approach: mostly dumb data
structs, and a circuit class that acts like a module to
orchestrate the simulation.

Prototyped a simple example of defining an op-amp circuit
with simple negative feedback using a pair of resistors.