Allocate forced voltage nodes only once.
There is no need to keep a separate device called FixedVoltage.
This just adds runtime overhead to the solver, for no benefit.
And in fact, it is less flexible, because there is no way for
outside code to control a variable forced voltage.

Instead, allocate a node as a forced voltage without setting a fixed voltage.
Return a reference to the floating point value of the node's voltage,
so that the outside caller can read/write it at will.

We are separating concepts: allocating a node as having a
forced voltage is a one-time configuration concept, whereas
setting the forced voltage to any given value can occur
as many times as desired during the simulation.

I should be able to use this same mechanism to implement
the output of an op-amp changing over time, with zero
overhead for solving the circuit.

The solver can treat forced nodes as having a given voltage,
and only search for solutions for non-forced nodes.

Also added convenience methods when we know for sure that
the node's voltage is not only forced but constant:

    setForcedVoltageNode
    setGroundNode