Fixed bug: do not return dangerous references!
When creating a component or a node, do not return a
reference to the thing we just added to a std::vector.
The problem is, the vector can get resized, which causes
the objects to be moved to a different location in memory,
leaving the reference pointing to freed memory.

Introduced the concept of "locking" the circuit.
You cannot access references to nodes or components
until the circuit is locked.
Once locked, you cannot add new nodes or components.
Any attempt to violate the rules results in an exception.
These rules prevent any possibility of accessing an
invalid reference later.