Got RC test to pass by using more reasonable component values.
Instead of using a 1 ohm resistor and a 1 farad capacitor,
use a 1M resistor and a 1uF capacitor. The time constant
is still the same: RC = 1 second.

But the current values are so much numerically smaller
(about a million times smaller) that it makes it easier
to get a current error on the order of 10 nA.
This makes me think the score should be computed in
dimensionless units, not in amps. If I can find
the maximum current moving through a given node,
I could scale by that amount.

The simulated capacitor ends up with a voltage error
of about 8 mV, which is pretty good.