Allow calling code to tweak the search parameters for a given circuit.
I'm realizing there is no easy way to automate picking
the best search parameters, especially the score tolerance.
So instead of hardcoding it, or trying to make a dimensionless
value that works in all circuits, just allow the caller to
set a member variable `scoreTolerance`. I did this in the
RC test and got much better results: 33uV for the final
capacitor error! That is well beyond what I need to emulate
a real-world circuit.