Creating separate programs for unit testing and animation.
I realized it won't work to have a single program for both
unit testing and animation.

The problem is I want to run the unit tests in GitHub Actions,
but I don't want to have to build something based on raylib there.
So the animation program needs to be separate, so I build it
locally only.