Fixed wrong node for `z` voltage, and renamed to `z`.
I had a mistake in an older schematic I made that used
node 8 as the LED output voltage, when it was actually
supposed to be node 7, according to the original plan
by Andrew Fitch.

Also, in the Triple Sloth, this voltage is labeled `z`.

Changed to use node 7, and renamed it from `LED` to `z`.